% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datarequests.R
\name{chs_data_request}
\alias{chs_data_request}
\title{Check data request}
\usage{
chs_data_request(
  data,
  case_list_path,
  case_list_name,
  output_path,
  output_name,
  ...
)
}
\arguments{
\item{data}{'data.frame' of Sentencing Guidelines public data set.}

\item{case_list_path}{pathway where the case list will be saved.}

\item{case_list_name}{name of the case list file.}

\item{output_path}{pathway where the report output will be saved.}

\item{output_name}{name of the report output file.}

\item{...}{represents any variable within the data set. When a variable is used
as an argument, the definition of the argument will be used to filter the data.
Four arguments that are derivatives of variables exist: variablename_min,
variablename_max, reason, and preason. variablename_min and _max are used for
selecting a range of data, (i.e., sentyear) and reason and preason are used for
selecting departure reasons and plea reasons.}
}
\description{
Check data request
}
