% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final-case-list-function.R
\name{comsetic_function}
\alias{comsetic_function}
\alias{case_list}
\title{Cosmetic function}
\usage{
case_list(df1)
}
\arguments{
\item{df1}{a filtered 'data.frame' of the cases requested in the data request.}
}
\value{
A filtered data frame with fewer variables.
}
\description{
Cosmetic function
}
\keyword{internal}
