% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/departure-and-plea-reasons-functions.R
\name{dep_reasons_function}
\alias{dep_reasons_function}
\alias{dep_reasons}
\title{Departure reasons function}
\usage{
dep_reasons(data_frame, dep_type = NULL)
}
\arguments{
\item{data_frame}{a filtered 'data.frame' of the departure reasons for
departures within the data set.}

\item{dep_type}{the type of departure to filter the data by: mit_disp, mit_dur,
agg_disp, or agg_dur.}
}
\value{
a 'data.frame' of the number of reasons and percent composition of departure
reasons for the data set.
}
\description{
Departure reasons function
}
\keyword{internal}
