% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-excel-function.R
\name{excel_function}
\alias{excel_function}
\alias{create_excel}
\title{Excel function}
\usage{
create_excel(data_set, name, path)
}
\arguments{
\item{data_set}{any 'data.frame'.}

\item{name}{name of the file.}

\item{path}{path where the file will be saved.}
}
\value{
a named excel file saved in the path provided and opened.
}
\description{
Excel function
}
\keyword{internal}
