% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non-public-cases-function.R
\name{non_public_function}
\alias{non_public_function}
\alias{non_public_cases}
\title{Non-public function}
\usage{
non_public_cases(data, filters)
}
\arguments{
\item{data}{a filtered 'data.frame' of the cases requested in the data request.}
}
\value{
The number of non-public cases.
}
\description{
Non-public function
}
\keyword{internal}
