% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chs-table-functions.R,
%   R/county-table-functions.R
\name{table_functions}
\alias{table_functions}
\alias{total_cases_by_chs}
\alias{pres_disp_cases}
\alias{disp_dep_cases}
\alias{dur_dep_cases}
\alias{prison_duration}
\alias{final_chs_table}
\alias{total_cases_by_county}
\alias{pres_disp_cases_by_county}
\alias{disp_dep_cases_by_county}
\alias{dur_dep_cases_by_county}
\alias{prison_duration_by_county}
\alias{final_table_by_county}
\title{Table functions}
\usage{
total_cases_by_chs(data_frame)

pres_disp_cases(data_frame)

disp_dep_cases(data_frame, pres_disp)

dur_dep_cases(data_frame)

prison_duration(data_frame)

final_chs_table(
  table_total_cases,
  table_pres_disp,
  table_disp_dep,
  table_dur_dep,
  table_pris_dur
)

total_cases_by_county(data_frame)

pres_disp_cases_by_county(data_frame)

disp_dep_cases_by_county(data_frame, pres_disp)

dur_dep_cases_by_county(data_frame)

prison_duration_by_county(data_frame)

final_table_by_county(
  table_total_cases,
  table_pres_disp,
  table_disp_dep,
  table_dur_dep,
  table_pris_dur
)
}
\arguments{
\item{data_frame}{a filtered 'data.frame' of the cases requested in the data request.}

\item{pres_disp}{'data.frame' of the total cases separated by presumptive disposition.}

\item{table_total_cases}{a 'data.frame' of the total cases for the data set grouped by county.}

\item{table_pres_disp}{a 'data.frame' of the total cases for the data set grouped by county
and presumptive disposition.}

\item{table_disp_dep}{a 'data.frame' of the total cases for the data set grouped by county
and dispositional departures. Percentages are calculated using presumptive
stays as the denominator for aggravated dispositional departures and presumptive
commits as the denominator for mitigated dispositional departures.}

\item{table_dur_dep}{a 'data.frame' of the total cases for the data set grouped by county
and durational departures. Percentages are calculated using prison sentences only.}

\item{table_pris_dur}{a 'data.frame' of the average prison duration sentenced for each
county.}
}
\value{
a 'data.frame' of the total cases for the data set grouped by Criminal
History Score.

a 'data.frame' of the total cases for the data set grouped by Criminal
History Score and presumptive disposition.

a 'data.frame' of the total cases for the data set grouped by Criminal
History Score and dispositional departures. Percentages are calculated using presumptive
stays as the denominator for aggravated dispositional departures and presumptive
commits as the denominator for mitigated dispositional departures.

a 'data.frame' of the total cases for the data set grouped by Criminal
History Score and durational departures. Percentages are calculated using
prison sentences only.

a 'data.frame' of the average prison duration sentenced for each Criminal History Score.

A 'data.frame' all data request table columns combined into one table.

a 'data.frame' of the total cases for the data set grouped by county.

a 'data.frame' of the total cases for the data set grouped by county and presumptive disposition.

a 'data.frame' of the total cases for the data set grouped by county
and dispositional departures. Percentages are calculated using presumptive
stays as the denominator for aggravated dispositional departures and presumptive
commits as the denominator for mitigated dispositional departures.

a 'data.frame' of the total cases for the data set grouped by county
and durational departures. Percentages are calculated using prison sentences only.

a 'data.frame' of the average prison duration sentenced for each county.

A 'data.frame' all data request table columns combined into one table.
}
\description{
Table functions

Table functions
}
\keyword{internal}
