% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_function}
\alias{filter_function}
\alias{requested_cases}
\title{Filter function}
\usage{
requested_cases(data, filters)
}
\arguments{
\item{data}{'data.frame' of Sentencing Guidelines public data set.}

\item{filters}{a named list of all arguments passed to the function.}
}
\value{
A filtered data frame.
}
\description{
Filter function
}
\keyword{internal}
