% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-functions.R
\name{table_functions}
\alias{table_functions}
\alias{total_cases_by_chs}
\alias{pres_disp_cases}
\alias{disp_dep_cases}
\alias{dur_dep_cases}
\alias{final_chs_table}
\title{Table functions}
\usage{
total_cases_by_chs(df1)

pres_disp_cases(data, df1)

disp_dep_cases(df1, pres_disp)

dur_dep_cases(df1)

final_chs_table(
  table_total_cases,
  table_pres_disp,
  table_disp_dep,
  table_dur_dep
)
}
\arguments{
\item{df1}{a filtered 'data.frame' of the cases requested in the data request.}

\item{pres_disp}{'data.frame' of the total cases separated by presumptive disposition.}

\item{table_total_cases}{a 'data.frame' of the total cases for the data set grouped by Criminal
History Score.}

\item{table_pres_disp}{a 'data.frame' of the total cases for the data set grouped by Criminal
History Score and presumptive disposition.}

\item{table_disp_dep}{a 'data.frame' of the total cases for the data set grouped by Criminal
History Score and dispositional departures. Percentages are calculated using presumptive
stays as the denominator for aggravated dispositional departures and presumptive
commits as the denominator for mitigated dispositional departures.}

\item{table_dur_dep}{a 'data.frame' of the total cases for the data set grouped by Criminal
History Score and durational departures. Percentages are calculated using
prison sentences only.}
}
\value{
a 'data.frame' of the total cases for the data set grouped by Criminal
History Score.

a 'data.frame' of the total cases for the data set grouped by Criminal
History Score and presumptive disposition.

a 'data.frame' of the total cases for the data set grouped by Criminal
History Score and dispositional departures. Percentages are calculated using presumptive
stays as the denominator for aggravated dispositional departures and presumptive
commits as the denominator for mitigated dispositional departures.

a 'data.frame' of the total cases for the data set grouped by Criminal
History Score and durational departures. Percentages are calculated using
prison sentences only.

A 'data.frame' all data request table columns combined into one table.
}
\description{
Table functions
}
\keyword{internal}
